add model choices file (#885)
* add model choices file

* fix imports for model choices

* fix rest of imports

---------
