+++ backend/samfundet/models/recruitment.py
@@ -11,6 +11,8 @@ from django.db import models
 from root.utils.mixins import FullCleanSaveMixin
 from .general import Organization, User, Gang
 
+from samfundet.models.model_choices import RecruitmentPriorityChoices, RecruitmentStatusChoices
+
 
 class Recruitment(FullCleanSaveMixin):
     name_nb = models.CharField(max_length=100, help_text='Name of the recruitment')
@@ -177,9 +179,11 @@ class RecruitmentAdmission(FullCleanSaveMixin):
     ]
 
     # TODO: Important that the following is not sent along with the rest of the object whenever a user retrieves its admission
-    recruiter_priority = models.IntegerField(choices=PRIORITY_CHOICES, default=0, help_text='The priority of the admission')
+    recruiter_priority = models.IntegerField(
+        choices=RecruitmentPriorityChoices, default=RecruitmentPriorityChoices.NOT_SET, help_text='The priority of the admission'
+    )
 
-    recruiter_status = models.IntegerField(choices=STATUS_CHOICES, default=0, help_text='The status of the admission')
+    recruiter_status = models.IntegerField(choices=RecruitmentStatusChoices, default=RecruitmentStatusChoices.NOT_SET, help_text='The status of the admission')
 
     def __str__(self) -> str:
         return f'Admission: {self.user} for {self.recruitment_position} in {self.recruitment}'
