+++ backend/samfundet/models/tests/test_billig.py
@@ -1,5 +1,6 @@
 from samfundet.models.billig import BilligEvent
-from samfundet.models.event import Event, EventTicketType
+from samfundet.models.event import Event
+from samfundet.models.model_choices import EventTicketType
 
 
 def test_billig_link_with_event(fixture_event: Event, fixture_billig_event: BilligEvent):
